+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added "Bonus to Clicking income" stat on clicking upgrades' tooltips
 * Added "Percentage of current income" in tooltips
 * Fixed Santa Upgrades position on the screen
+* Prevent tooltips from going out of the screen
 
 ### 1.040.11 _(2014-01-05)_